1.0

+++ config.hpp
@@ -2,11 +2,12 @@
 #define TINYFM_CONFIG_HPP
 
 #include "stdint.h"
+#include "fixed.hpp"
 
 typedef fixed32<16> TimeType;
 typedef fixed32<16> SampleType;
 
-const float freq = 8000.0f;
+const int freq = 8000;
 
 #define PI 3.141592f
